// packages/directus-modules/hook-sync-to-legacy/src/btv.d.ts
export interface components {
  schemas: {
// ...
    ItemsShows: {
      agerating_code?:
        | (string | components["schemas"]["ItemsAgeratings"])
        | null;
      available_from?: string | null;
      available_to?: string | null;
      date_created?: string;
      date_updated?: string;
      id?: number;
      image_file_id?: (string | components["schemas"]["Files"]) | null;
      legacy_description_id?: number | null;
      legacy_id?: number | null;
      legacy_title_id?: number | null;
      publish_date?: string;
      status?: string;
      type?: string;
      user_created?: (string | components["schemas"]["Users"]) | null;
      user_updated?: (string | components["schemas"]["Users"]) | null;
      default_episode_behaviour?: string | null;
      /** Should the episode production date be visible in title */
      publish_date_in_title?: boolean | null;
      /** For internal use only */
      label?: string | null;
      public_title?: string | null;
      images?: (number | components["schemas"]["ItemsImages"])[];
      availability?: string;
      seasons?: (number | components["schemas"]["ItemsSeasons"])[];
      usergroups?: (number | components["schemas"]["ItemsShowsUsergroups"])[];
      translations?: (
        | number
        | components["schemas"]["ItemsShowsTranslations"]
      )[];
      metadata?: string;
      related?: string;
      tags?: (number | components["schemas"]["ItemsShowsTags"])[];
      technical_details?: string;
    };
// ...
    Files: {
      /** Unique identifier for the file. */
      id?: string;
      /** Where the file is stored. Either `local` for the local filesystem or the name of the storage adapter (for example `s3`). */
      storage?: string;
      /** Name of the file on disk. By default, Directus uses a random hash for the filename. */
      filename_disk?: string;
      /** How you want to the file to be named when it's being downloaded. */
      filename_download?: string;
      /** Title for the file. Is extracted from the filename on upload, but can be edited by the user. */
      title?: string;
      /** MIME type of the file. */
      type?: string;
      /** Virtual folder where this file resides in. */
      folder?: (string | components["schemas"]["Folders"]) | null;
      /** Who uploaded the file. */
      uploaded_by?: string | components["schemas"]["Users"];
      /** When the file was uploaded. */
      uploaded_on?: string;
      modified_by?: (string | components["schemas"]["Users"]) | null;
      modified_on?: string;
      /** Character set of the file. */
      charset?: string | null;
      /** Size of the file in bytes. */
      filesize?: number;
      /** Width of the file in pixels. Only applies to images. */
      width?: number | null;
      /** Height of the file in pixels. Only applies to images. */
      height?: number | null;
      /** Duration of the file in seconds. Only applies to audio and video. */
      duration?: number | null;
      /** Where the file was embedded from. */
      embed?: string | null;
      /** Description for the file. */
      description?: string | null;
      /** Where the file was created. Is automatically populated based on EXIF data for images. */
      location?: string | null;
      /** Tags for the file. Is automatically populated based on EXIF data for images. */
      tags?: string[] | null;
      /** IPTC, EXIF, and ICC metadata extracted from file */
      metadata?: { [key: string]: any } | null;
      storage_divider?: string;
    };
// ...
  };
// ...
}